requirements.txt
Flask==3.0.0
gunicorn==21.2.0

numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2

requests==2.31.0
python-dotenv==1.0.1

google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

logging

